Fix form builder.
